# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 知卷原创命题核心业务 API（设计契约）
  version: 1.3.0
  description: 知卷拟开发业务接口；不是 Hermes 原生 API，也没有随文档提供运行服务器。v1.3 补齐当前蓝图读取、完整规格读取、历史列表、当前身份与发布撤回，并将运行对账状态显式化。端到端和生产身份实现仍待开发。
  x-zhijuan-runtime:
    agent: hermes
    provider: deepseek
    additional-agent-frameworks: []
servers:
- url: http://localhost:8000
  description: 仅本地开发示例，生产必须 HTTPS。
security:
- session: []
paths:
...
  /v1/exams:
    post:
      operationId: createExam
      summary: 创建试卷草稿
      description: ''
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSummary'
          headers:
            ETag:
              schema:
                type: string
                minLength: 1
                maxLength: 100
              description: 当前 revision
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '412':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamSpec'
    get:
      operationId: listExams
      summary: 分页读取本组织可访问的试卷项目
      description: 只提供项目元数据，不实现跨卷题目检索。按创建时间与 id 稳定排序，游标绑定租户和过滤条件。
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: cursor
        in: query
        schema:
          type: string
          maxLength: 1000
      responses:
        '200':
          description: 成功（拟开发）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamList'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/exams/{exam_id}:
    get:
      operationId: getExam
      summary: 按权限返回学生或教师投影
      description: teacher 视图必须具备读取答案权限；student 响应中不包含 private 字段。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ExamStudentView'
                - $ref: '#/components/schemas/ExamTeacherView'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '412':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
      - name: exam_id
        in: path
        required: true
        schema: &id001
          type: string
          format: uuid
      - name: view
        in: query
        schema:
          type: string
          enum:
          - student
          - teacher
          default: student
  /v1/exams/{exam_id}/spec:
    patch:
      operationId: replaceExamSpec
      summary: 替换完整配置并使计划失效
      description: ''
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSummary'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '412':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
      - name: exam_id
        in: path
        required: true
        schema: *id001
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamSpec'
    get:
      operationId: getExamSpec
      summary: 读取当前完整命题配置
      description: 授权后返回完整规格与 ETag。用于页面恢复，不从 Agent 历史重建。
      parameters:
      - name: exam_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: 成功（拟开发）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSpec'
          headers:
            ETag:
              schema:
                type: string
              description: 带引号的当前 exam revision；修改请求用 If-Match 回传。
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /v1/exams/{exam_id}/generation-jobs:
    post:
      operationId: generateExam
      summary: 按已确认蓝图生成试卷
      description: 必须有与当前规格一致且已确认的计划；无计划/未确认/已过期时拒绝。正常失败、取消或未知上游结果不能直接重发计费请求。
      responses:
        '202':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '412':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
      - name: exam_id
        in: path
        required: true
        schema: *id001
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
...
  /v1/exams/{exam_id}/plans/{plan_id}/confirm:
    post:
      operationId: confirmPlan
      summary: 确认指定版本蓝图
      description: ''
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSummary'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '412':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
      - name: exam_id
        in: path
        required: true
        schema: *id001
      - name: plan_id
        in: path
        required: true
        schema: *id001
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                expected_revision:
                  type: integer
                  minimum: 1
                  maximum: 100000
                plan_hash: &id002
                  type: string
                  pattern: ^[0-9a-f]{64}$
              required:
              - expected_revision
              - plan_hash
...
  /v1/exams/{exam_id}/plans/current:
    get:
      operationId: getCurrentPlan
      summary: 读取待确认或已确认的当前命题蓝图
      description: 无 current_plan_id 返回 404 PLAN_NOT_FOUND；与当前规格不符返回 409 PLAN_STALE。只返回该卷的当前计划，不携带答案。确认接口继续使用 plan_id、plan_hash
        和期望 exam revision。
      parameters:
      - name: exam_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: 成功（拟开发）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentPlan'
          headers:
            ETag:
              schema:
                type: string
              description: 当前 exam revision 的强 ETag；plan_hash 是单独的计划内容指纹。
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  schemas:
    ExamSpec:
      $ref: ./exam-spec.schema.json
    Candidate:
      $ref: ./candidate.schema.json
    PublicQuestion:
      $ref: ./public-question.schema.json
...
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 80
        message:
          type: string
          minLength: 1
          maxLength: 2000
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
          minLength: 1
          maxLength: 128
        retryable:
          type: boolean
      required:
      - code
      - message
      - details
      - trace_id
      - retryable
    ExamSummary:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        revision:
          type: integer
          minimum: 1
          maximum: 100000
        state:
          type: string
          enum:
          - DRAFT
          - PLANNING
          - PLAN_READY
          - GENERATING
          - REVIEW_READY
          - APPROVED
          - PUBLISHED
          - ARCHIVED
        title:
          type: string
          minLength: 1
          maxLength: 120
        content_hash:
          type:
          - string
          - 'null'
          pattern: ^[0-9a-f]{64}$
      required:
      - id
      - revision
      - state
      - title
      - content_hash
    JobAccepted:
      type: object
      additionalProperties: false
      properties:
        job_id: *id001
        status:
          type: string
          enum:
          - QUEUED
        exam_revision:
          type: integer
          minimum: 1
          maximum: 100000
      required:
      - job_id
      - status
      - exam_revision
...
    JobRequest:
      type: object
      additionalProperties: false
      properties:
        expected_revision:
          type: integer
          minimum: 1
          maximum: 100000
        budget_microunits:
          type: integer
          minimum: 1
          maximum: 1000000000000
        currency:
          type: string
          minLength: 1
          maxLength: 8
        note:
          type: string
          minLength: 0
          maxLength: 2000
      required:
      - expected_revision
      - budget_microunits
      - currency
      - note
...
    ExamStudentView:
      type: object
      additionalProperties: false
      properties:
        exam:
          $ref: '#/components/schemas/ExamSummary'
        view:
          const: student
        questions:
          type: array
          items:
            $ref: '#/components/schemas/PublicQuestion'
          minItems: 0
          maxItems: 1000
      required:
      - exam
      - view
      - questions
    ExamTeacherView:
      type: object
      additionalProperties: false
      properties:
        exam:
          $ref: '#/components/schemas/ExamSummary'
        view:
          const: teacher
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
          minItems: 0
          maxItems: 1000
      required:
      - exam
      - view
      - candidates
...
    PlanSlot:
      type: object
      additionalProperties: false
      properties:
        slot_id: *id004
        score_x100:
          type: integer
          minimum: 0
        knowledge_ids:
          type: array
          items: *id004
          minItems: 1
        objective:
          type: string
          minLength: 1
          maxLength: 3000
        constraints:
          type: array
          items:
            type: string
        material_group_id:
          type:
          - string
          - 'null'
          maxLength: 120
        kind:
          type: string
          enum:
          - single_choice
          - multiple_choice
          - true_false
          - fill_blank
          - solution
          - short_answer
          - essay
          - material_group
        parent_slot_id:
          type:
          - string
          - 'null'
          minLength: 1
          maxLength: 200
          description: 根槽位为 null；子题引用本计划 material_group 槽位。服务端检查存在性、无环及最大深度。
      required:
      - slot_id
      - kind
      - score_x100
      - knowledge_ids
      - objective
      - constraints
      - material_group_id
      - parent_slot_id
      allOf:
      - if:
          properties:
            kind:
              const: material_group
          required:
          - kind
        then:
          properties:
            score_x100:
              const: 0
        else:
          properties:
            score_x100:
              minimum: 1
      description: 含材料父槽位和叶子题。材料父槽位分值固定为0，只累加叶子题；parent_slot_id形成受校验的无环依赖。
    CurrentPlan:
      type: object
      additionalProperties: false
      properties:
        plan_id: *id003
        exam_id: *id003
        spec_revision: &id005
          type: integer
          minimum: 1
        plan_revision: *id005
        plan_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
        slots:
          type: array
          items:
            $ref: '#/components/schemas/PlanSlot'
          minItems: 1
          maxItems: 1000
        confirmed_by:
          type:
          - string
          - 'null'
          format: uuid
        confirmed_at:
          type:
          - string
          - 'null'
          format: date-time
      required:
      - plan_id
      - exam_id
      - spec_revision
      - plan_revision
      - plan_hash
      - slots
      - confirmed_by
      - confirmed_at
    ExamList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExamSummary'
          maxItems: 100
        next_cursor:
          type:
          - string
          - 'null'
          maxLength: 1000
      required:
      - items
      - next_cursor
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: zhijuan_session
      description: 同源 HTTPS 会话；所有写请求另需 X-CSRF-Token。Agent 内部鉴权不使用此 Cookie。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
      description: 同一键不同请求体返回409；保留窗口需在部署策略中冻结。
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 100
      description: 例如 "exam-7"；缺失428，不匹配412。
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 256
  responses:
    Error:
      description: 统一错误。查看 code；不回显密钥、内部路径或敏感正文。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...